moves inspector content to workspace
