+++ web-local/src/routes/(application)/dashboard/[id]/+page.svelte
@@ -17,4 +17,12 @@
   <title>Rill Developer</title>
 </svelte:head>
 
+<!-- 
+  temporary workaround to clear out non-white bg space. 
+  once the q4 application refactor is complete let's remove this.
+-->
+<!-- <div
+  class="bg-white absolute left-0 top-0 right-0 bottom-0 w-screen h-screen"
+  style:border="4px solid black"
+/> -->
 <Explore {metricsDefId} />
